requirements.txt
bcrypt==3.1.7
cffi==1.12.3
Django==1.10
django-multiselectfield==0.1.8
pycparser==2.19
pytz==2019.1
six==1.12.0
